Rename tool_use/search to search_tool
Move the search tool recipe from tinker_cookbook/recipes/tool_use/search/
to tinker_cookbook/recipes/search_tool/ and update all internal imports.
