+++ tinker_cookbook/recipes/search_tool/search_env.py
@@ -14,7 +14,7 @@ import tinker
 from huggingface_hub import hf_hub_download
 from tinker_cookbook import renderers
 from tinker_cookbook.completers import StopCondition
-from tinker_cookbook.recipes.tool_use.search.tools import ChromaToolClient, ChromaToolClientConfig
+from tinker_cookbook.recipes.search_tool.tools import ChromaToolClient, ChromaToolClientConfig
 from tinker_cookbook.rl.problem_env import ProblemEnv, ProblemGroupBuilder
 from tinker_cookbook.rl.types import (
     Action,
